Move generic_tasks and processing to helpers
